# docs/swagger.yaml
basePath: /api
definitions:
...
  github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_admin_request.UpdateCommentStatus:
    properties:
      id:
        minimum: 1
        type: integer
      status:
        enum:
        - approved
        - rejected
        - spam
        type: string
    required:
    - id
    - status
    type: object
  github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_admin_request.UpdateFeedbackStatus:
    properties:
      id:
        minimum: 1
        type: integer
      status:
        enum:
        - pending
        - processing
        - resolved
        - closed
        type: string
    required:
    - id
    - status
    type: object
...
  github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_admin_response.CommentDetail:
    properties:
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      ip_address:
        type: string
      likes:
        type: integer
      parent_id:
        type: integer
      post_id:
        type: integer
      replies_count:
        type: integer
      status:
        type: string
      updated_at:
        type: string
      user_agent:
        type: string
      user_id:
        type: integer
      user_profile:
        $ref: '#/definitions/github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_admin_response.UserProfile'
    type: object
  github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_admin_response.CommentDetailPage:
    properties:
      current_page:
        type: integer
      list:
        items:
          $ref: '#/definitions/github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_admin_response.CommentDetail'
        type: array
      page_size:
        type: integer
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
...
  github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_admin_response.FeedbackDetail:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      ip_address:
        type: string
      message:
        type: string
      name:
        type: string
      status:
        type: string
      subject:
        type: string
      type:
        type: string
      updated_at:
        type: string
      user_agent:
        type: string
    type: object
  github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_admin_response.FeedbackDetailPage:
    properties:
      current_page:
        type: integer
      list:
        items:
          $ref: '#/definitions/github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_admin_response.FeedbackDetail'
        type: array
      page_size:
        type: integer
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
...
  github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_admin_response.UserProfile:
    properties:
      avatar:
        type: string
      bio:
        type: string
      blog_url:
        type: string
      email:
        type: string
      name:
        type: string
      region:
        type: string
      status:
        type: string
    type: object
  github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_shared.Envelope:
    properties:
      code:
        example: "0000"
        type: string
      data: {}
      message:
        example: ok
        type: string
    type: object
...
info:
  contact: {}
  description: Xiaoqing Blog 后端 API（Admin + Public V1）。
  title: Xiaoqing Blog API
  version: "1.0"
paths:
...
  /admin/comments/:
    get:
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 分页大小
        in: query
        name: page_size
        type: integer
      - description: 排序字段
        enum:
        - created_at
        in: query
        name: sort_by
        type: string
      - default: desc
        description: 排序方向
        enum:
        - asc
        - desc
        in: query
        name: order
        type: string
      - description: 状态过滤
        in: query
        name: filter.status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_shared.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_admin_response.CommentDetailPage'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_shared.Envelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_shared.Envelope'
      security:
      - AdminSession: []
      summary: 评论列表
      tags:
      - Admin.Comments
  /admin/comments/{id}:
    delete:
      parameters:
      - description: 评论 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_shared.Envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_shared.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_shared.Envelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_shared.Envelope'
      security:
      - AdminSession: []
      summary: 删除评论
      tags:
      - Admin.Comments
  /admin/comments/{id}/status:
    put:
      consumes:
      - application/json
      parameters:
      - description: 评论 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 状态信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_admin_request.UpdateCommentStatus'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_shared.Envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_shared.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_shared.Envelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_shared.Envelope'
      security:
      - AdminSession: []
      summary: 更新评论状态
      tags:
      - Admin.Comments
...
  /admin/feedbacks/:
    get:
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 分页大小
        in: query
        name: page_size
        type: integer
      - description: 排序字段
        enum:
        - created_at
        in: query
        name: sort_by
        type: string
      - default: desc
        description: 排序方向
        enum:
        - asc
        - desc
        in: query
        name: order
        type: string
      - description: 状态过滤
        in: query
        name: filter.status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_shared.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_admin_response.FeedbackDetailPage'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_shared.Envelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_shared.Envelope'
      security:
      - AdminSession: []
      summary: 反馈列表
      tags:
      - Admin.Feedbacks
  /admin/feedbacks/{id}:
    delete:
      parameters:
      - description: 反馈 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_shared.Envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_shared.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_shared.Envelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_shared.Envelope'
      security:
      - AdminSession: []
      summary: 删除反馈
      tags:
      - Admin.Feedbacks
    get:
      parameters:
      - description: 反馈 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_shared.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_admin_response.FeedbackDetail'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_shared.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_shared.Envelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_shared.Envelope'
      security:
      - AdminSession: []
      summary: 反馈详情
      tags:
      - Admin.Feedbacks
  /admin/feedbacks/{id}/status:
    put:
      consumes:
      - application/json
      parameters:
      - description: 反馈 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 状态信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_admin_request.UpdateFeedbackStatus'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_shared.Envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_shared.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_shared.Envelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_QING520-MAKER_Xiaoqing-blog-api_internal_controller_http_shared.Envelope'
      security:
      - AdminSession: []
      summary: 更新反馈状态
      tags:
      - Admin.Feedbacks
...
schemes:
- http
- https
securityDefinitions:
  AdminSession:
    in: cookie
    name: fiber_session
    type: apiKey
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"